# api/openapi.yaml
openapi: 3.0.3
info:
  title: ClubHub API
  description: |
    ClubHub 高校社团运营与协同管理平台 API。

    **API-first 开发流程**：修改本文件 → push → CI 自动生成前后端代码 →
    git pull 拉取生成代码 → 在 Controller / Services / Vue 组件中手写业务逻辑。
  version: 0.1.0
servers:
  - url: http://localhost:5000
    description: 本地开发服务器
paths:
...
  /api/auth/register:
    post:
      summary: 用户注册
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: 注册信息不合法
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: 用户名或学工号已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/auth/login:
    post:
      summary: 用户登录
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功，返回用户和角色信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: 登录信息不合法
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: 用户名/学工号或密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: 账号已被禁用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/auth/session:
    get:
      summary: 刷新当前用户会话
      description: 依据 Bearer 登录令牌中的当前用户重新计算角色、社团身份和权限，用于成员身份变化后的前端会话同步。
      operationId: refreshAuthSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 已重新计算的用户会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: 登录状态缺失或令牌无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: 账号已被禁用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/auth/permissions/check:
    get:
      summary: 检查用户是否拥有指定权限
      operationId: checkPermission
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: permission
          in: query
          required: true
          schema:
            type: string
        - name: clubId
          in: query
          required: false
          description: Optional club id filter.
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: 权限判断结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionCheckResult"
        "400":
          description: 请求参数不合法
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/auth/roles/assign:
    post:
      summary: 分配用户角色
      operationId: assignUserRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRoleRequest"
      responses:
        "200":
          description: 角色分配成功或已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignmentResult"
        "400":
          description: 请求参数不合法
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: 操作人没有角色分配权限
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: 用户或角色不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ClubHub-HMAC
      description: 登录或注册成功后返回的 ClubHub Bearer 令牌。
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
          nullable: true
          example: venue_reservation_conflict
        message:
          type: string
          example: 该场地在所选时间段已有已通过预约。
        detail:
          type: string
          nullable: true
      required:
        - message
...
    RegisterRequest:
      type: object
      description: 用户注册请求。学生学工号为 7 位数字，教师工号为 5 位数字。
      properties:
        username:
          type: string
          description: 登录用户名，注册后必须唯一。
          minLength: 3
          maxLength: 50
          example: stu_public
        password:
          type: string
          description: 登录密码，服务端仅保存密码摘要。
          minLength: 6
          maxLength: 128
          example: ClubHub123
        realName:
          type: string
          description: 用户真实姓名。
          minLength: 1
          maxLength: 50
          example: 张三
        studentNo:
          type: string
          description: 学工号；学生 7 位，教师 5 位。
          minLength: 5
          maxLength: 7
          pattern: "^([0-9]{7}|[0-9]{5})$"
          example: "2450001"
        gender:
          type: string
          description: 性别，可为空；当前支持男、女。
          nullable: true
          maxLength: 10
          pattern: "^(男|女)$"
          example: 男
        phone:
          type: string
          description: 联系电话，可为空。
          nullable: true
          maxLength: 20
          pattern: "^[0-9+\\-\\s()]{0,20}$"
          example: "13800138000"
        email:
          type: string
          description: 邮箱地址，可为空。
          nullable: true
          format: email
          maxLength: 100
          example: zhangsan@example.com
        college:
          type: string
          description: 所属学院，可为空。
          nullable: true
          maxLength: 100
          example: 软件学院
        major:
          type: string
          description: 所属专业，可为空。
          nullable: true
          maxLength: 100
          example: 软件工程
        grade:
          type: string
          description: 年级，可为空。
          nullable: true
          maxLength: 20
          example: "2024"
      required:
        - username
        - password
        - realName
        - studentNo
    LoginRequest:
      type: object
      description: 用户登录请求，username 可填写用户名或学工号。
      properties:
        username:
          type: string
          description: 用户名或学工号。
          minLength: 1
          maxLength: 50
          example: stu_public
        password:
          type: string
          description: 登录密码。
          minLength: 1
          maxLength: 128
          example: ClubHub123
      required:
        - username
        - password
    AssignRoleRequest:
      type: object
      description: 为目标用户分配系统级或社团级角色。
      properties:
        operatorUserId:
          type: integer
          description: 发起分配操作的用户 ID。
          example: 1
        targetUserId:
          type: integer
          description: 被分配角色的用户 ID。
          example: 2
        roleCode:
          type: string
          description: 角色编码，例如 STUDENT、CLUB_MEMBER、SYSTEM_ADMIN。
          maxLength: 50
          example: CLUB_MEMBER
        clubId:
          type: integer
          description: 社团级角色所属社团 ID；系统级角色可为空。
          nullable: true
          example: 1
      required:
        - operatorUserId
        - targetUserId
        - roleCode
    AuthResponse:
      type: object
      description: 注册或登录成功后的认证响应。
      properties:
        token:
          type: string
          description: 登录令牌；调用需要当前登录态的接口时通过 Authorization Bearer 头传递。
          example: eyJ1c2VySWQiOjF9.abc123signature
        user:
          description: 当前登录用户信息。
          $ref: "#/components/schemas/AuthUser"
        roles:
          type: array
          description: 当前用户拥有的角色列表。
          items:
            $ref: "#/components/schemas/AuthRole"
        permissions:
          type: array
          description: 当前用户角色权限的并集。
          items:
            type: string
          example: ["profile:view", "public:view"]
      required:
        - token
        - user
        - roles
        - permissions
    AuthUser:
      type: object
      description: 认证上下文中的用户基础信息。
      properties:
        id:
          type: integer
          description: 用户 ID。
          example: 1
        username:
          type: string
          description: 登录用户名。
          example: stu_public
        realName:
          type: string
          description: 用户真实姓名。
          example: 张三
        studentNo:
          type: string
          description: 学工号；学生 7 位，教师 5 位。
          nullable: true
          example: "2450001"
        gender:
          type: string
          description: 性别。
          nullable: true
          example: 男
        phone:
          type: string
          description: 联系电话。
          nullable: true
          example: "13800138000"
        email:
          type: string
          description: 邮箱地址。
          nullable: true
          format: email
          example: zhangsan@example.com
        college:
          type: string
          description: 所属学院。
          nullable: true
          example: 软件学院
        major:
          type: string
          description: 所属专业。
          nullable: true
          example: 软件工程
        grade:
          type: string
          description: 年级。
          nullable: true
          example: "2024"
        accountStatus:
          type: string
          description: 账号状态，例如 normal 或 disabled。
          example: normal
      required:
        - id
        - username
        - realName
        - accountStatus
    AuthRole:
      type: object
      description: 用户在认证上下文中的角色及其权限。
      properties:
        id:
          type: integer
          description: 角色 ID。
          example: 3
        code:
          type: string
          description: 角色编码。
          example: CLUB_MEMBER
        name:
          type: string
          description: 角色名称。
          example: 社团成员
        displayName:
          type: string
          description: 前端展示用角色名称，例如“A社团成员”“A社团负责人”。
          example: 计算机协会成员
        scope:
          type: string
          description: 角色作用域，system 表示全局，club 表示社团范围。
          enum: [system, club]
          example: club
        clubId:
          type: integer
          description: 当前角色对应的单个社团 ID；全局角色为空。
          nullable: true
          example: 1
        clubIds:
          type: array
          description: 该角色覆盖的社团范围；用于展示和权限判断中的社团范围角色。
          items:
            type: integer
          example: [1]
        permissions:
          type: array
          description: 该角色授予的权限编码列表。
          items:
            type: string
          example: ["club:internal:view", "activity:checkin"]
        permissionDesc:
          type: string
          description: 角色权限说明。
          nullable: true
          example: 可查看社团内部信息并参与活动签到。
      required:
        - id
        - code
        - name
        - displayName
        - scope
        - clubIds
        - permissions
...
    PermissionCheckResult:
      type: object
      description: 权限检查结果。
      properties:
        userId:
          type: integer
          description: 被检查的用户 ID。
          example: 1
        permission:
          type: string
          description: 被检查的权限编码。
          example: club:internal:view
        clubId:
          type: integer
          description: 权限检查涉及的社团 ID；全局权限可为空。
          nullable: true
          example: 1
        allowed:
          type: boolean
          description: 是否允许访问。
          example: true
        matchedRoles:
          type: array
          description: 命中该权限的角色列表。
          items:
            $ref: "#/components/schemas/AuthRole"
        message:
          type: string
          description: 面向调用方的判定说明。
          example: 允许访问。
      required:
        - userId
        - permission
        - allowed
        - matchedRoles
        - message
    RoleAssignmentResult:
      type: object
      description: 角色分配结果。
      properties:
        targetUserId:
          type: integer
          description: 被分配角色的用户 ID。
          example: 2
        role:
          description: 已分配或已存在的角色。
          $ref: "#/components/schemas/AuthRole"
        alreadyExists:
          type: boolean
          description: 该角色分配是否原本已存在。
          example: false
        message:
          type: string
          description: 角色分配结果说明。
          example: 角色分配成功。
      required:
        - targetUserId
        - role
        - alreadyExists
        - message
...